Relocates watcher testing infrastructure to package_test.go, and into the cache package from cache_test.
